More OOP 90 percent implemented
